Add plot_splits_hist for polar histograms
`plot_splits_hist` and `plot_splits_hist!` by default use a
different colour for each station, but this can be overridden
by passing a formatting keyword argument through to Plots.

Also add `bar` to create a bar `Shape` which can be used in future
to replace the quiver plot for `plot_splits_bar`.